chore[php]: re-org functions
