+++ index.php
@@ -11,43 +11,7 @@
         $page = "home";
     }
 
-    function Redirect($url, $permanent = false) {
-        header('Location: ' . $url, true, $permanent ? 301 : 302);
-        exit();
-    }
-
-    function BrowserRedirect($url)
-    {
-        $string = '<script type="text/javascript">';
-        $string .= 'window.location = "' . $url . '"';
-        $string .= '</script>';
-
-        echo $string;
-        exit();
-    }
-
-    /**
-     * Function: sanitize
-     * Returns a sanitized string, typically for URLs.
-     *
-     * Parameters:
-     *     $string - The string to sanitize.
-     *     $force_lowercase - Force the string to lowercase?
-     *     $anal - If set to *true*, will remove all non-alphanumeric characters.
-     */
-    function sanitize($string, $force_lowercase = true, $anal = false) {
-        $strip = array("~", "`", "!", "@", "#", "$", "%", "^", "&", "*", "(", ")", "_", "=", "+", "[", "{", "]",
-                    "}", "\\", "|", ";", ":", "\"", "'", "&#8216;", "&#8217;", "&#8220;", "&#8221;", "&#8211;", "&#8212;",
-                    "â€”", "â€“", ",", "<", ".", ">", "/", "?");
-        $clean = trim(str_replace($strip, "", strip_tags($string)));
-        $clean = preg_replace('/\s+/', "_", $clean);
-        $clean = ($anal) ? preg_replace("/[^a-zA-Z0-9]/", "", $clean) : $clean ;
-        return ($force_lowercase) ?
-            (function_exists('mb_strtolower')) ?
-                mb_strtolower($clean, 'UTF-8') :
-                strtolower($clean) :
-            $clean;
-    }
+    require 'modules/functions.php';
 
     if (!empty($_POST['redirect'])) {
         $redirect = $_POST['redirect'];
