+++ php/functions.php
@@ -1,0 +1,49 @@
+<?php
+
+/**
+ * Function: sanitize
+ * Returns a sanitized string, typically for URLs.
+ *
+ * Parameters:
+ *     $string - The string to sanitize.
+ *     $force_lowercase - Force the string to lowercase?
+ *     $anal - If set to *true*, will remove all non-alphanumeric characters.
+ */
+function sanitize($string, $force_lowercase = true, $anal = false) {
+    $strip = array("~", "`", "!", "@", "#", "$", "%", "^", "&", "*", "(", ")", "_", "=", "+", "[", "{", "]",
+                "}", "\\", "|", ";", ":", "\"", "'", "&#8216;", "&#8217;", "&#8220;", "&#8221;", "&#8211;", "&#8212;",
+                "â€”", "â€“", ",", "<", ".", ">", "/", "?");
+    $clean = trim(str_replace($strip, "", strip_tags($string)));
+    $clean = preg_replace('/\s+/', "_", $clean);
+    $clean = ($anal) ? preg_replace("/[^a-zA-Z0-9]/", "", $clean) : $clean ;
+    return ($force_lowercase) ?
+        (function_exists('mb_strtolower')) ?
+            mb_strtolower($clean, 'UTF-8') :
+            strtolower($clean) :
+        $clean;
+}
+
+function Redirect($url, $permanent = false) {
+    header('Location: ' . $url, true, $permanent ? 301 : 302);
+    exit();
+}
+
+function BrowserRedirect($url) {
+    $string = '<script type="text/javascript">';
+    $string .= 'window.location = "' . $url . '"';
+    $string .= '</script>';
+
+    echo $string;
+    exit();
+}
+
+function set_progress_file($operation) {
+    if (file_exists($progress_file)) { unlink($progress_file); }
+    $progress = array(
+        "operation" => $operation,
+        "completed" => "-",
+        "total" => "-"
+    );
+    $json = json_encode($progress, JSON_UNESCAPED_SLASHES);
+    file_put_contents($progress_file, $json);
+}
